create new ImageTracertest for image tracer tests
